Add exceptions to the routes

+++ app.py
@@ -9,23 +9,30 @@ app = Flask(__name__)
 @db_session
 def create_book():
     data = request.json
-    name = data['name']
+    name = data['name'].strip().lower()
     author = data['author']
 
-    Livro(name=name, author=author)
-    return jsonify(
-        {"Status": "Book Created"})
+    book = Livro.get(name = name)
+    if book:
+        return jsonify({"Status": "Book already exists"})
+    else:
+        Livro(name=name, author=author)
+        return jsonify({"Status": "Book Created"})
+        
 
 #Rota para Ler um livro de acordo com seu id
 @app.route('/<string:id>', methods=['GET'])
 @db_session
 def get_byid(id):
     book = Livro.get(id=id)
+    if not book:
+        return jsonify({"Status": "Book not found"})
     
-    return jsonify(
+    else:
+        return jsonify(
         {"Status": "Response",
         "Id": book.id,
-        "Name": book.name,
+        "Name": book.name.strip().title(),
         "Author": book.author}
         )
     
@@ -35,26 +42,30 @@ def get_byid(id):
 @db_session
 def update_byid(id):
     data = request.json
-    name = data['name']
+    name = data['name'].strip().lower()
     author = data['author']
     book = Livro.get(id=id)
     
-    book.name = name
-    book.author = author
-
-    return jsonify({
-        "Status": "Book Updated"
-    })
+    if book:
+        book.name = name
+        book.author = author
+        return jsonify({
+        "Status": "Book Updated"})
+    else:
+        return jsonify({"Status": "Book not found"})
 
 #Rota para deletar livro de acordo com seu id.
 @app.route('/<string:id>', methods=['DELETE'])
 @db_session
 def delete_byid(id):
     book = Livro.get(id=id)
-    book.delete()
-    return jsonify({
-        "Status": "Book Deleted"
-    })
+    if not book:
+        return jsonify({"Status": "Book not found"})
+    
+    else:
+        book.delete()
+        return jsonify({
+        "Status": "Book Deleted"})
 
 
 #Rota para retornar todos os livros do banco de dados.
@@ -62,7 +73,7 @@ def delete_byid(id):
 @db_session
 def get_allbooks():
     all_books = Livro.select()[:]
-    livros_list = [{"id": str(livro.id), "name": livro.name, "author": livro.author} for livro in all_books]
+    livros_list = [{"id": str(livro.id), "name": livro.name.strip().title(), "author": livro.author} for livro in all_books]
     return jsonify(livros_list)
